Actually this is the fastq aggregation collation script
